+++ Demo/Sources/GlimpseDemoApp.swift
@@ -1,0 +1,17 @@
+//
+//  GlimpseDemoApp.swift
+//  GlimpseDemo
+//
+//  Created by Nazar Kozak on 12.06.2026.
+//
+
+import SwiftUI
+
+@main
+struct GlimpseDemoApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}